# guangdong-energy-monitor/docs/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: 广东省电厂监控平台 API 契约
  description: |
    华电集团广东电厂监控平台后端接口契约文档。
    所有接口统一通过 Gateway（8080）访问，需携带 JWT Token。
    所有响应统一使用 R<T> 包装结构。
    包含认证授权、用户管理、角色管理、组织架构、电厂管理、水文站管理、
    GIS 空间服务、报警管理、预警管理、规则管理、统计分析、预测服务、数据采集共13个控制器、54个REST端点。
  version: 1.0.0
  contact:
    name: 华电集团广东电厂监控平台研发组
servers:
  - url: http://localhost:8080
    description: 本地开发环境（Gateway 入口）
  - url: https://api.gd-energy.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/gis/plants/bounds:
    get:
      tags: [GIS 空间服务]
      summary: 按范围查询电厂（GeoJSON）
      description: 在指定经纬度范围内查询电厂，返回 GeoJSON FeatureCollection
      parameters:
        - name: minLng
          in: query
          required: true
          description: 最小经度
          schema:
            type: number
            format: double
        - name: maxLng
          in: query
          required: true
          description: 最大经度
          schema:
            type: number
            format: double
        - name: minLat
          in: query
          required: true
          description: 最小纬度
          schema:
            type: number
            format: double
        - name: maxLat
          in: query
          required: true
          description: 最大纬度
          schema:
            type: number
            format: double
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantFeatureCollection'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/gis/hydro/stations/bounds:
    get:
      tags: [GIS 空间服务]
      summary: 按范围查询水文站（GeoJSON）
      description: 在指定经纬度范围内查询水文站，返回 GeoJSON FeatureCollection
      parameters:
        - name: minLng
          in: query
          required: true
          description: 最小经度
          schema:
            type: number
            format: double
        - name: maxLng
          in: query
          required: true
          description: 最大经度
          schema:
            type: number
            format: double
        - name: minLat
          in: query
          required: true
          description: 最小纬度
          schema:
            type: number
            format: double
        - name: maxLat
          in: query
          required: true
          description: 最大纬度
          schema:
            type: number
            format: double
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HydroStationFeatureCollection'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/gis/distance:
    get:
      tags: [GIS 空间服务]
      summary: 计算两点距离
      description: 计算两个经纬度坐标之间的距离
      parameters:
        - name: lng1
          in: query
          required: true
          description: 第一个点的经度
          schema:
            type: number
            format: double
        - name: lat1
          in: query
          required: true
          description: 第一个点的纬度
          schema:
            type: number
            format: double
        - name: lng2
          in: query
          required: true
          description: 第二个点的经度
          schema:
            type: number
            format: double
        - name: lat2
          in: query
          required: true
          description: 第二个点的纬度
          schema:
            type: number
            format: double
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: number
                        format: double
                        description: 两点之间距离（公里）
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # 报警管理 (AlarmController) - 4 endpoints
  # ============================================================
...
  /api/stats/dashboard:
    get:
      tags: [统计分析]
      summary: 首页统计看板数据
      description: 返回首页大屏所需的汇总统计：电厂总数、各状态分布、报警统计、预警统计等
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/stats/plant-trend:
    get:
      tags: [统计分析]
      summary: 电厂趋势分析
      description: 返回电厂负荷/出力的历史趋势数据（时间序列格式）
      parameters:
        - name: plantId
          in: query
          required: true
          description: 电厂ID
          schema:
            type: integer
            format: int64
        - name: metric
          in: query
          required: true
          description: 指标类型
          schema:
            type: string
            enum: [load, output, capacity]
        - name: startTime
          in: query
          required: true
          description: 开始时间
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          description: 结束时间
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PlantTrendVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/stats/alarm-statistics:
    get:
      tags: [统计分析]
      summary: 报警统计分析
      description: 按时间范围统计报警数据
      parameters:
        - name: startTime
          in: query
          required: true
          description: 开始时间
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          description: 结束时间
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/StatItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/stats/hydro-statistics:
    get:
      tags: [统计分析]
      summary: 水文统计分析
      description: 按水文站和时间范围统计水文数据
      parameters:
        - name: stationId
          in: query
          required: true
          description: 水文站ID
          schema:
            type: integer
            format: int64
        - name: startTime
          in: query
          required: true
          description: 开始时间
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          description: 结束时间
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          additionalProperties:
                            type: object
                          description: 水文统计项（Map结构）
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # 预测服务 (PredictController) - 2 endpoints
  # ============================================================
...
  /api/collector/weather/trigger:
    get:
      tags: [数据采集]
      summary: 触发气象数据采集
      description: 手动触发气象数据采集任务，返回最新采集的气象数据列表
      responses:
        '200':
          description: 采集成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/WeatherData'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/collector/hydro/trigger:
    get:
      tags: [数据采集]
      summary: 触发水文数据采集
      description: 手动触发病文数据采集任务，返回最新采集的水文数据列表
      responses:
        '200':
          description: 采集成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HydroData'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/collector/emergency/trigger:
    get:
      tags: [数据采集]
      summary: 触发应急事件采集
      description: 手动触发应急事件数据采集任务，返回最新采集的应急事件列表
      responses:
        '200':
          description: 采集成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EmergencyEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/collector/weather/{plantId}:
    get:
      tags: [数据采集]
      summary: 查询电厂气象数据
      description: 查询指定电厂的最新气象数据
      parameters:
        - name: plantId
          in: path
          required: true
          description: 电厂ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WeatherData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/collector/hydro/{stationId}/readings:
    get:
      tags: [数据采集]
      summary: 查询水文站读数
      description: 查询指定水文站在时间范围内的历史读数数据
      parameters:
        - name: stationId
          in: path
          required: true
          description: 水文站ID
          schema:
            type: integer
            format: int64
        - name: startTime
          in: query
          required: true
          description: 开始时间
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          description: 结束时间
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HydroData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

# ================================================================
# Components
# ================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token 认证，登录后获取 token，请求头携带 Authorization: Bearer <token>"
...
  responses:
...
    Unauthorized:
      description: 未认证或Token无效（401 UNAUTHORIZED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
    NotFound:
      description: 资源不存在（404 NOT_FOUND）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
  schemas:
...
    R:
      type: object
      description: |
        统一响应包装结构 R<T>。
        错误码：200 SUCCESS, 400 PARAM_ERROR, 401 UNAUTHORIZED,
        402 TOKEN_EXPIRED, 403 FORBIDDEN, 404 NOT_FOUND,
        500 BUSINESS_ERROR, 5000 SYSTEM_ERROR
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        msg:
          type: string
          description: 提示信息
          example: 操作成功
        data:
          description: 业务数据（具体结构见各接口）
          nullable: true

    # --------------------------------------------------------
    # 分页结果 PageResult<T>
    # --------------------------------------------------------
...
    WeatherData:
      type: object
      description: 气象数据
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        plantId:
          type: integer
          format: int64
          description: 电厂ID
        temp:
          type: number
          format: double
          description: 温度（℃）
        humidity:
          type: number
          format: double
          description: 湿度（%）
        windSpeed:
          type: number
          format: double
          description: 风速（m/s）
        windDirection:
          type: string
          description: 风向
        rainfall:
          type: number
          format: double
          description: 降雨量（mm）
        recordTime:
          type: string
          format: date-time
          description: 记录时间

    # --------------------------------------------------------
    # 水文数据
    # --------------------------------------------------------
    HydroData:
      type: object
      description: 水文数据
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        stationId:
          type: integer
          format: int64
          description: 水文站ID
        waterLevel:
          type: number
          format: double
          description: 水位（m）
        flowRate:
          type: number
          format: double
          description: 流量（m³/s）
        trend:
          type: string
          description: 水位趋势
          enum: [rising, falling, steady]
        alertLevel:
          type: string
          description: 告警等级
          enum: [normal, watch, warning, flood]
        readingTime:
          type: string
          format: date-time
          description: 读数时间

    # --------------------------------------------------------
    # 应急事件
    # --------------------------------------------------------
    EmergencyEvent:
      type: object
      description: 应急事件
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        type:
          type: string
          description: 事件类型
        level:
          type: string
          description: 事件等级
        title:
          type: string
          description: 事件标题
        content:
          type: string
          description: 事件内容
        location:
          type: string
          description: 事件地点
        lng:
          type: number
          format: double
          description: 经度
        lat:
          type: number
          format: double
          description: 纬度
        occurTime:
          type: string
          format: date-time
          description: 发生时间
        status:
          type: string
          description: 状态

    # --------------------------------------------------------
    # 仪表盘
    # --------------------------------------------------------
    DashboardVO:
      type: object
      description: 首页统计看板视图对象
      properties:
        totalPlants:
          type: integer
          description: 电厂总数
        normalCount:
          type: integer
          description: 正常电厂数
        warningCount:
          type: integer
          description: 预警电厂数
        dangerCount:
          type: integer
          description: 危险电厂数
        offlineCount:
          type: integer
          description: 离线电厂数
        totalAlarms:
          type: integer
          description: 报警总数
        activeWarnings:
          type: integer
          description: 活跃预警数
        hydroStationCount:
          type: integer
          description: 水文站总数
        hydroAlertCount:
          type: integer
          description: 水文告警数

    # --------------------------------------------------------
    # 趋势分析
    # --------------------------------------------------------
    PlantTrendVO:
      type: object
      description: 电厂趋势分析（时间序列格式）
      properties:
        plantId:
          type: integer
          format: int64
          description: 电厂ID
        plantName:
          type: string
          description: 电厂名称
        dates:
          type: array
          description: 时间序列
          items:
            type: string
            format: date-time
        values:
          type: array
          description: 数值序列
          items:
            type: number
            format: double
        metric:
          type: string
          description: 指标类型

    # --------------------------------------------------------
    # 统计项
    # --------------------------------------------------------
    StatItem:
      type: object
      description: 统计项
      properties:
        label:
          type: string
          description: 标签
        value:
          type: number
          format: double
          description: 数值
        color:
          type: string
          description: 显示颜色

    # --------------------------------------------------------
    # 预测响应
    # --------------------------------------------------------
...
    PlantType:
      type: string
      description: 电厂类型
      enum:
        - coal
        - gas
        - solar
        - wind
        - storage
    PlantStatus:
      type: string
      description: 电厂运行状态
      enum:
        - normal
        - warning
        - danger
        - offline
...
    GeoJSONFeatureCollection:
      type: object
      description: GeoJSON FeatureCollection 基础结构
      required:
        - type
        - features
      properties:
        type:
          type: string
          enum: [FeatureCollection]
        features:
          type: array
          items:
            $ref: '#/components/schemas/GeoJSONFeature'
    GeoJSONFeature:
      type: object
      description: GeoJSON Feature
      required:
        - type
        - geometry
        - properties
      properties:
        type:
          type: string
          enum: [Feature]
        geometry:
          $ref: '#/components/schemas/GeoJSONGeometry'
        properties:
          type: object
          additionalProperties: true
    GeoJSONGeometry:
      type: object
      description: GeoJSON Geometry
      required:
        - type
        - coordinates
      properties:
        type:
          type: string
          enum: [Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon]
        coordinates:
          type: array
          items: {}
    PlantFeatureCollection:
      allOf:
        - $ref: '#/components/schemas/GeoJSONFeatureCollection'
        - type: object
          description: 电厂 GeoJSON FeatureCollection，properties 包含 Plant 全部字段
          properties:
            features:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/GeoJSONFeature'
                  - type: object
                    properties:
                      properties:
                        type: object
                        properties:
                          id:
                            type: integer
                            format: int64
                          name:
                            type: string
                          type:
                            $ref: '#/components/schemas/PlantType'
                          lng:
                            type: number
                            format: double
                          lat:
                            type: number
                            format: double
                          capacity:
                            type: number
                            format: double
                          status:
                            $ref: '#/components/schemas/PlantStatus'
                          address:
                            type: string
    HydroStationFeatureCollection:
      allOf:
        - $ref: '#/components/schemas/GeoJSONFeatureCollection'
        - type: object
          description: 水文站 GeoJSON FeatureCollection，properties 包含 HydroStation 全部字段
          properties:
            features:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/GeoJSONFeature'
                  - type: object
                    properties:
                      properties:
                        type: object
                        properties:
                          id:
                            type: integer
                            format: int64
                          name:
                            type: string
                          riverBasin:
                            type: string
                          lng:
                            type: number
                            format: double
                          lat:
                            type: number
                            format: double
                          warningLevel:
                            type: number
                            format: double
                          guaranteeLevel:
                            type: number
                            format: double
                          historicalMax:
                            type: number
                            format: double
                          flowCapacity:
                            type: number
                            format: double
                          city:
                            type: string
                          status:
                            type: integer